Moved redis mock client to separate file
